requirements.txt
flask
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
groq